@@ -39,25 +39,25 @@
     "relaks-event-emitter": "2.0.8"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.7",
-    "@babel/plugin-transform-runtime": "^7.8.3",
-    "@babel/preset-env": "^7.8.7",
-    "@babel/runtime": "^7.8.7",
-    "@rollup/plugin-node-resolve": "^7.1.1",
-    "babel-loader": "^8.0.6",
+    "@babel/core": "^7.10.2",
+    "@babel/plugin-transform-runtime": "^7.10.1",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/runtime": "^7.10.2",
+    "@rollup/plugin-node-resolve": "^8.0.0",
+    "babel-loader": "^8.1.0",
     "chai": "^4.2.0",
-    "karma": "^4.4.1",
+    "karma": "^5.0.9",
     "karma-babel-preprocessor": "^8.0.1",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-mocha": "^1.3.0",
+    "karma-mocha": "^2.0.1",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.32",
     "karma-webpack": "^4.0.2",
     "lodash": "^4.17.15",
-    "mocha": "^7.1.1",
-    "rollup": "^2.1.0",
+    "mocha": "^7.2.0",
+    "rollup": "^2.13.1",
     "rollup-plugin-babel": "^4.4.0",
-    "webpack": "^4.42.0"
+    "webpack": "^4.43.0"
   }
 }
